+++ backend/src/main/java/com/cogumelos/dto/formulacao/FormulacaoRequest.java
@@ -24,5 +24,5 @@ public record FormulacaoRequest(
         @NotNull @Size(min=1) List<FormulacaoInsumoItem> insumos,
         @NotNull @DecimalMin(value = "0.1", message = "Umidade desejada deve ser maior que zero.") Double umidadeDesejada,
         @NotNull @DecimalMin(value = "0.1", message = "Peso por bloco deve ser maior que zero.") Double pesoBlocoKg,
-        Integer totalBlocos
+        Double totalBlocos
 ) {}
